Fix slider path anchor coordinates being parsed to `int` instead of `float`
osu!lazer exports them to `float`s, so the current behavior will invalidate such sliders.

+++ src/com/rian/osu/beatmap/parser/sections/BeatmapHitObjectsParser.kt
@@ -108,10 +108,7 @@ object BeatmapHitObjectsParser : BeatmapSectionParser() {
                     scope?.ensureActive()
                     it.isEmpty()
                 }.let {
-                    val curvePointPosition = Vector2(
-                        parseInt(it[0]).toFloat(),
-                        parseInt(it[1]).toFloat()
-                    )
+                    val curvePointPosition = Vector2(parseFloat(it[0]), parseFloat(it[1]))
 
                     curvePoints.add(curvePointPosition - startPosition)
                 }
